refactor: move ToMfxCodec to vpl.[hc]pp
